requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.14
django-crispy-forms==1.13.0
django-gravatar2==1.4.4
django-hCaptcha==0.1.0
django-utils-six==2.0
django-widget-tweaks==1.4.9
idna==3.3
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.4.0
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.2.0
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.8.6
django-imap-backend==0.2.3
dj-database-url==0.5.0
psycopg2-binary
dropbox
pillow
